Version 1.8.0

+++ common/sonar.spec
@@ -1,75 +1,37 @@
 ################################################################################
 
-# rpmbuilder:relative-pack true
-
-################################################################################
-
 %global crc_check pushd ../SOURCES ; sha512sum -c %{SOURCE100} ; popd
 
 ################################################################################
 
-%define _posixroot        /
-%define _root             /root
-%define _bin              /bin
-%define _sbin             /sbin
-%define _srv              /srv
-%define _home             /home
-%define _opt              /opt
-%define _lib32            %{_posixroot}lib
-%define _lib64            %{_posixroot}lib64
-%define _libdir32         %{_prefix}%{_lib32}
-%define _libdir64         %{_prefix}%{_lib64}
-%define _logdir           %{_localstatedir}/log
-%define _rundir           %{_localstatedir}/run
-%define _lockdir          %{_localstatedir}/lock/subsys
-%define _cachedir         %{_localstatedir}/cache
-%define _spooldir         %{_localstatedir}/spool
-%define _crondir          %{_sysconfdir}/cron.d
-%define _loc_prefix       %{_prefix}/local
-%define _loc_exec_prefix  %{_loc_prefix}
-%define _loc_bindir       %{_loc_exec_prefix}/bin
-%define _loc_libdir       %{_loc_exec_prefix}/%{_lib}
-%define _loc_libdir32     %{_loc_exec_prefix}/%{_lib32}
-%define _loc_libdir64     %{_loc_exec_prefix}/%{_lib64}
-%define _loc_libexecdir   %{_loc_exec_prefix}/libexec
-%define _loc_sbindir      %{_loc_exec_prefix}/sbin
-%define _loc_bindir       %{_loc_exec_prefix}/bin
-%define _loc_datarootdir  %{_loc_prefix}/share
-%define _loc_includedir   %{_loc_prefix}/include
-%define _loc_mandir       %{_loc_datarootdir}/man
-%define _rpmstatedir      %{_sharedstatedir}/rpm-state
-%define _pkgconfigdir     %{_libdir}/pkgconfig
+%define debug_package  %{nil}
 
 ################################################################################
 
-%define debug_package     %{nil}
+%define _logdir  %{_localstatedir}/log
 
 ################################################################################
 
-%define srcdir            src/github.com/essentialkaos/%{name}
+Summary:        Utility for showing user Slack status in Atlassian Jira
+Name:           sonar
+Version:        1.8.0
+Release:        0%{?dist}
+Group:          Applications/System
+License:        Apache License, Version 2.0
+URL:            https://kaos.sh/sonar
 
-################################################################################
+Source0:        https://source.kaos.st/%{name}/%{name}-%{version}.tar.bz2
 
-Summary:         Utility for showing user Slack status in Atlassian Jira
-Name:            sonar
-Version:         1.7.1
-Release:         0%{?dist}
-Group:           Applications/System
-License:         Apache License, Version 2.0
-URL:             https://github.com/essentialkaos/sonar
+Source100:      checksum.sha512
 
-Source0:         https://source.kaos.st/%{name}/%{name}-%{version}.tar.bz2
+BuildRoot:      %{_tmppath}/%{name}-%{version}-%{release}-root-%(%{__id_u} -n)
 
-Source100:       checksum.sha512
+BuildRequires:  golang >= 1.19
 
-BuildRoot:       %{_tmppath}/%{name}-%{version}-%{release}-root-%(%{__id_u} -n)
+Requires:       kaosv >= 2.16
+Requires:       systemd
 
-BuildRequires:   golang >= 1.14
-
-Requires:        kaosv >= 2.16
-Requires:        systemd
-
-Provides:        %{name} = %{version}-%{release}
+Provides:       %{name} = %{version}-%{release}
 
 ################################################################################
 
@@ -84,35 +46,38 @@ Utility for showing user Slack status in Atlassian Jira.
 %setup -q
 
 %build
-export GOPATH=$(pwd)
+if [[ ! -d "%{name}/vendor" ]] ; then
+  echo "This package requires vendored dependencies"
+  exit 1
+fi
 
-pushd %{srcdir}
-  %{__make} %{?_smp_mflags} sonar
+pushd %{name}
+  %{__make} %{?_smp_mflags} all
+  cp LICENSE ..
 popd
 
 %install
 rm -rf %{buildroot}
 
-install -dm 755 %{buildroot}%{_bindir}
-install -dm 755 %{buildroot}%{_sysconfdir}
-install -dm 755 %{buildroot}%{_sysconfdir}/logrotate.d
-install -dm 755 %{buildroot}%{_initddir}
-install -dm 755 %{buildroot}%{_unitdir}
-install -dm 755 %{buildroot}%{_logdir}/%{name}
+install -dDm 755 %{buildroot}%{_bindir}
+install -dDm 755 %{buildroot}%{_sysconfdir}/logrotate.d
+install -dDm 755 %{buildroot}%{_initddir}
+install -dDm 755 %{buildroot}%{_unitdir}
+install -dDm 755 %{buildroot}%{_logdir}/%{name}
 
-install -pm 755 %{srcdir}/%{name} \
+install -pm 755 %{name}/%{name} \
                 %{buildroot}%{_bindir}/
 
-install -pm 644 %{srcdir}/common/%{name}.knf \
+install -pm 644 %{name}/common/%{name}.knf \
                 %{buildroot}%{_sysconfdir}/
 
-install -pm 755 %{srcdir}/common/%{name}.init \
+install -pm 755 %{name}/common/%{name}.init \
                 %{buildroot}%{_initddir}/%{name}
 
-install -pm 644 %{srcdir}/common/%{name}.logrotate \
+install -pm 644 %{name}/common/%{name}.logrotate \
                 %{buildroot}%{_sysconfdir}/logrotate.d/%{name}
 
-install -pDm 644 %{srcdir}/common/%{name}.service \
+install -pDm 644 %{name}/common/%{name}.service \
                  %{buildroot}%{_unitdir}/
 
 %clean
@@ -138,6 +103,11 @@ exit 0
 ################################################################################
 
 %changelog
+* Thu Mar 31 2022 Anton Novojilov <andy@essentialkaos.com> - 1.8.0-0
+- Removed pkg.re usage
+- Smaller, optimized SVG images
+- Fixed major bug with checking token
+
 * Thu Sep 10 2020 Anton Novojilov <andy@essentialkaos.com> - 1.7.1-0
 - Fixed compatibility with the latest version of ek package
